Improve Sync interface

+++ test/v1/test_account_info.py
@@ -21,8 +21,12 @@ import six
 
 from .test_base import TestBase
 
-from .deps import AbstractAccountInfo, InMemoryAccountInfo, UploadUrlPool, SqliteAccountInfo
+from .deps import AbstractAccountInfo, InMemoryAccountInfo, UploadUrlPool
 from .deps_exception import CorruptAccountInfo, MissingAccountData
+
+if not platform.system().lower().startswith('java'):
+    # in Jython 2.7.1b3 there is no sqlite3
+    from .deps import SqliteAccountInfo
 
 try:
     import unittest.mock as mock
@@ -212,6 +216,9 @@ class TestSqliteAccountInfo(AccountInfoBase, TestBase):
         ).name
 
     def setUp(self):
+        if platform.system().lower().startswith('java'):
+            # in Jython 2.7.1b3 there is no sqlite3
+            raise SkipTest()
         try:
             os.unlink(self.db_path)
         except OSError:
